chore: tidying up builds without datafusion feature and clippy
The recent convert_to_delta changes imported a symbol from a module that
is only compiled in with the datafusion feature, oops!
